+++ theory/tests/test_periodic.c
@@ -30,10 +30,10 @@ void read_data_and_set_globals(const char *firstfilename, const char *firstforma
 
 
 //Global variables
-int ND1;
+int64_t ND1;
 double *X1=NULL,*Y1=NULL,*Z1=NULL,*weights1=NULL;
 
-int ND2;
+int64_t ND2;
 double *X2=NULL,*Y2=NULL,*Z2=NULL,*weights2=NULL;
 
 char current_file1[MAXLEN],current_file2[MAXLEN];
